Fix make fifo

+++ gis-weather.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
 #
 #  gis_weather.py
-v = '0.8.2.37'
+v = '0.8.2.38'
 #  Copyright (C) 2013-2017 Alexander Koltsov <ringov@mail.ru>
 #
 #    This program is free software: you can redistribute it and/or modify
@@ -888,11 +888,13 @@ class MyDrawArea(Gtk.DrawingArea):
             if os.path.exists(save_cur_data_path):
                 path_to_save = save_cur_data_path
             if save_cur_temp_to_pipe:
-                try:
-                    os.mkfifo(os.path.join(path_to_save, 'cur_temp'))
-                except:
+                if os.path.isfile(os.path.join(path_to_save, 'cur_temp')):
                     os.remove(os.path.join(path_to_save, 'cur_temp'))
                     os.mkfifo(os.path.join(path_to_save, 'cur_temp'))
+                if not os.path.exists(os.path.join(path_to_save, 'cur_temp')):
+                    os.mkfifo(os.path.join(path_to_save, 'cur_temp'))
+            elif not os.path.isfile(os.path.join(path_to_save, 'cur_temp')) and os.path.exists(os.path.join(path_to_save, 'cur_temp')):
+                os.remove(os.path.join(path_to_save, 'cur_temp'))
             process = subprocess.Popen(['/bin/bash'], shell=False, stdin=subprocess.PIPE, stdout=subprocess.PIPE, bufsize=1, universal_newlines=True)
             process.stdin.write('echo "'+t_now[0].split(';')[t_index]+t_now_post+'" > '+os.path.join(path_to_save, 'cur_temp')+'\n')
             print(os.path.join(path_to_save, 'cur_temp')+' saved (%s)'%t_now[0].split(';')[t_index]+t_now_post)
